Fix various security issues in the JS client and the demo code
The demo code has these three problems:

- Signature validation failure doesn't cause an early exit.
- The string the HMAC is validated on is not the string the user sent.
- HMACs are not compared in constant time.

You could reasonably argue that this doesn't really matter because (1) it's
demo code (2) doesn't reveal any information to the attacker that they
didn't know.

However, in cryptography, (2) is generally still considered a problem
because it's bad to have multiple valid representations of the same
message. For instance, various naive ways of preventing replay attacks
will fall over if the attacker can generate multiple valid messages that
are regarded as having the same HMAC.

As for (3), it's academically considered a problem because a smart
client can usually brute force the authentication system.

There's a few platform-specific problems -- I fixed the Python ones,
but not the Node ones:

- Your code would flunk Pyright because you assign None to a str
  value in some dead code.
- metriport.ts passes `req.body` (a deserialized JSON value) to
  verifyWebhookSignature (which expects a string, but converts
  the value to bytes anyway)

Do not merge this without testing it yourself -- this is in draft
status because I do not have a metriport dev environment set up. I am
simply including the change to show what needs to be fixed.

Sorry for the trouble -- good luck, your product seems great!